scripts/requirements.txt
# MKTG Skills Research Agent Dependencies

# HTTP client
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.0

# CLI enhancements (optional)
rich>=13.0.0
typer>=0.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
